requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-dotenv==1.0.1
orjson==3.10.7
httpx==0.27.2
pytest==8.3.3
SQLAlchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
python-multipart>=0.0.6
aiofiles>=23.2.0
pandas>=2.1.0

